AI inf less likely to die due to colliding with other AI inf
Fun fact: the pattern in which assets  spawn around you for a given # isn't random.

If AI inf are still dying when you call in 8 inf at once...its possible we can't fix this since at that point the pattern might be doubling over on itself and literally spawn two guys in the same place. I can't seem to find anywhere to undo this repeating pattern thing.