RespawnCounter error fixed (#237)
This was the root of most of the spawning related errors
- Spawning in water fixed
- spawning at 0,0 fixed
- respawning where you died fixed
- Cant teleport error fixed

Likely other stuff got fixed too, this was basically bricking the mission when it happened.

What cause the error: New sectors were added a few updates ago and some of them were too small so when they were selected as a base the init respawn code would error out. The init spawn error then lead to all the downstream errors.

solution: All zones that can be a base are now at least 150 x 150 meters in size. If any new sectors are added in the future make them at least 150 x 150 or add canBeBase false code to the sector init.